Feat/v2

+++ src/main/java/com/moyeoit/domain/review/controller/response/ReviewLikeResponse.java
@@ -2,6 +2,6 @@ package com.moyeoit.domain.review.controller.response;
 
 public record ReviewLikeResponse(
         boolean liked,  // 현재 내가 좋아요 눌렀는지
-        int likeCount
+        Long likeCount
 ) {
 }
